move admin test to admin.rs
